Put implementation away from widgets
