+++ lib/src/widgets/dashboard_screen/dashboard_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart' hide AppBar, Card;
 import 'package:saraka/constants.dart';
 import 'package:saraka/widgets.dart';
 import './app_bar.dart';
-import './main_drawer.dart';
+import './main_drawer/main_drawer.dart';
 import './proficient_card_prediction_text.dart';
 import './ready_card_length_text.dart';
 import './start_learning_floating_action_button.dart';
