+++ lib/src/widgets/signed_out_screen/privacy_policy_text.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/widgets.dart';
+import 'package:provider/provider.dart';
+import 'package:saraka/blocs.dart';
+import 'package:saraka/constants.dart';
+import 'package:saraka/widgets.dart';
+
+class PrivacyPolicyItem extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) => TappableText(
+        "Privacy Policy",
+        style: SarakaTextStyles.body,
+        onTap: () => Provider.of<CommonLinkBloc>(context).launchPrivacyPolicy(),
+      );
+}
