+++ lib/src/widgets/signed_out_screen/signed_out_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:saraka/constants.dart';
-import 'package:saraka/widgets.dart';
+import './privacy_policy_text.dart';
 import './sign_in_button.dart';
 
 class SignedOutScreen extends StatelessWidget {
@@ -41,11 +41,7 @@ class SignedOutScreen extends StatelessWidget {
                   SizedBox(height: 64),
                   SignInButton(),
                   SizedBox(height: 16),
-                  UrlLaunchableText(
-                    "Privacy Policy",
-                    url: privacyPolicyUrl,
-                    style: SarakaTextStyles.body,
-                  ),
+                  PrivacyPolicyItem(),
                 ],
               ),
             ),
