+++ lib/widgets.dart
@@ -22,5 +22,5 @@ export './src/widgets/signed_in_navigator/signed_in_navigator.dart';
 export './src/widgets/signed_out_screen/signed_out_screen.dart';
 export './src/widgets/review_screen/review_screen.dart';
 export './src/widgets/dashboard_screen/dashboard_screen.dart';
-export './src/widgets/url_launchable_text.dart';
+export './src/widgets/tappable_text.dart';
 export './src/widgets/wave_background.dart';
